+++ Source/lib/embeddedLapack/src/Lapack/Scr/dlasq1.c
@@ -27,7 +27,7 @@ static integer c__0 = 0;
     floatreal d__1, d__2, d__3;
 
     /* Builtin functions */
-    float sqrt(floatreal);
+    float sqrtf(floatreal);
 
     /* Local variables */
     integer i__;
@@ -180,7 +180,7 @@ static integer c__0 = 0;
 
     eps = dlamch_("Precision");
     safmin = dlamch_("Safe minimum");
-    scale = sqrt(eps / safmin);
+    scale = sqrtf(eps / safmin);
     dcopy_(n, &d__[1], &c__1, &work[1], &c__2);
     i__1 = *n - 1;
     dcopy_(&i__1, &e[1], &c__1, &work[2], &c__2);
@@ -205,7 +205,7 @@ static integer c__0 = 0;
     if (*info == 0) {
 	i__1 = *n;
 	for (i__ = 1; i__ <= i__1; ++i__) {
-	    d__[i__] = sqrt(work[i__]);
+	    d__[i__] = sqrtf(work[i__]);
 /* L40: */
 	}
 	dlascl_("G", &c__0, &c__0, &scale, &sigmx, n, &c__1, &d__[1], n, &
