+++ Source/lib/embeddedLapack/src/Lapack/Scr/dlasq3.c
@@ -25,7 +25,7 @@
     floatreal d__1, d__2;
 
     /* Builtin functions */
-    float sqrt(floatreal);
+    float sqrtf(floatreal);
 
     /* Local variables */
     floatreal s, t;
@@ -190,9 +190,9 @@ L40:
 	t = (z__[nn - 7] - z__[nn - 3] + z__[nn - 5]) * .5;
 	s = z__[nn - 3] * (z__[nn - 5] / t);
 	if (s <= t) {
-	    s = z__[nn - 3] * (z__[nn - 5] / (t * (sqrt(s / t + 1.) + 1.)));
+	    s = z__[nn - 3] * (z__[nn - 5] / (t * (sqrtf(s / t + 1.) + 1.)));
 	} else {
-	    s = z__[nn - 3] * (z__[nn - 5] / (t + sqrt(t) * sqrt(t + s)));
+	    s = z__[nn - 3] * (z__[nn - 5] / (t + sqrtf(t) * sqrtf(t + s)));
 	}
 	t = z__[nn - 7] + (s + z__[nn - 5]);
 	z__[nn - 3] *= z__[nn - 7] / t;
