+++ Source/lib/embeddedLapack/src/Lapack/Scr/sgeev.c
@@ -29,7 +29,7 @@ static integer c_n1 = -1;
     real r__1, r__2;
 
     /* Builtin functions */
-    float sqrt(floatreal);
+    float sqrtf(floatreal);
 
     /* Local variables */
     integer i__, k;
@@ -325,7 +325,7 @@ static integer c_n1 = -1;
     smlnum = slamch_("S");
     bignum = 1.f / smlnum;
     slabad_(&smlnum, &bignum);
-    smlnum = sqrt(smlnum) / eps;
+    smlnum = sqrtf(smlnum) / eps;
     bignum = 1.f / smlnum;
 
 /*     Scale A if max element outside range [SMLNUM,BIGNUM] */
