+++ Source/lib/embeddedLapack/src/Lapack/Scr/slapy2.c
@@ -19,7 +19,7 @@ floatreal slapy2_(real *x, real *y)
     real ret_val, r__1;
 
     /* Builtin functions */
-    float sqrt(floatreal);
+    float sqrtf(floatreal);
 
     /* Local variables */
     real w, z__, xabs, yabs;
@@ -35,7 +35,7 @@ floatreal slapy2_(real *x, real *y)
 /*  Purpose */
 /*  ======= */
 
-/*  SLAPY2 returns sqrt(x**2+y**2), taking care not to cause unnecessary */
+/*  SLAPY2 returns sqrtf(x**2+y**2), taking care not to cause unnecessary */
 /*  overflow. */
 
 /*  Arguments */
@@ -64,7 +64,7 @@ floatreal slapy2_(real *x, real *y)
     } else {
 /* Computing 2nd power */
 	r__1 = z__ / w;
-	ret_val = w * sqrt(r__1 * r__1 + 1.f);
+	ret_val = w * sqrtf(r__1 * r__1 + 1.f);
     }
     return ret_val;
